Port the tree to modern MSVC with a CMake build
The upstream source release targets Visual C++ 6.0 and a set of proprietary
SDKs that were never distributed with it, so it does not build as shipped.
This makes it build, and run, against nothing but Visual Studio and the
Windows SDK.

Build system:

- CMakeLists.txt replaces Code/commando.dsw. MSBuild cannot read .dsp/.dsw
  at all, so this is a replacement rather than a conversion. The .dsp files
  stay the source of truth for source lists; cmake/gen_sources.cmake reads
  them into cmake/Sources.cmake, run by hand with cmake -P.
- Compiler settings reproduce the VC6 semantics the code depends on:
  /Zc:forScope-, /Zc:wchar_t-, _USE_32BIT_TIME_T, /permissive. Each is
  commented where it is set.
- 32-bit only. 35 files contain x86 inline assembly.

Third-party dependencies, none of which now need downloading:

- DirectX 8 -> Direct3D 9, whose headers ship in the Windows SDK.
  Direct3DCreate9 is resolved at runtime, so no import library is needed.
  Code/Stubs/d3dx supplies the few D3DX entry points the engine uses.
- Miles Sound System -> Code/Stubs/miles6 implements the used API surface
  on XAudio2 and X3DAudio, which the engine drives without knowing the
  difference. A silent stub remains available as a fallback.
- Bink -> stubbed. BinkOpen fails, which the movie player already treats as
  "movie finished", so intros are skipped rather than hung.
- GameSpy -> stubbed. The services shut down in 2014; the used surface is
  four qr_* and five gcd_* calls. Costs the server browser and CD-key
  validation, not the game.
- GNU regex is absent from the release. regexpr.cpp is self-contained and
  unreferenced, so it is the one file left in cmake/Unported.cmake.

Also:

- Disc-check code (Code/Commando/cdverify.*) removed.
- BandTest links statically instead of as a DLL; it exports one function
  and its DllMain just returns true.
- Code/Tests/unit adds a suite over wwmath and wwlib, the two libraries
  everything else stands on and the only ones with no device or file
  dependencies. Run with ctest.
- README rewritten for this fork.

+++ Code/BandTest/BandTest.cpp
@@ -784,7 +784,10 @@ unsigned long Upstream_Detect(unsigned long server_ip, unsigned long my_ip, int 
 	*/
 	unsigned long downstream_bandwidth = upstream_bandwidth;
 	int old_band = Get_Registry_Int("Up", 0);
-	unsigned long diff = abs(upstream_bandwidth - old_band);
+	// The operand is unsigned long, so abs() is ambiguous under modern C++.
+	// VC6 had only abs(int) and narrowed to it silently; the cast preserves
+	// that exact behaviour on this 32-bit target.
+	unsigned long diff = abs((long)(upstream_bandwidth - old_band));
 	bool calc_down = true;
 	if (diff < upstream_bandwidth / 10) {
 		downstream_bandwidth = Get_Registry_Int("Down", upstream_bandwidth);
